Fix track floating percentage

+++ src/main/java/cam72cam/immersiverailroading/tile/TileRailBase.java
@@ -618,8 +618,11 @@ public class TileRailBase extends SyncdTileEntity implements ITrack, ITickable {
 				return;
 			}
 			
-			if (this instanceof TileRail) {
+			if (Config.ConfigDamage.requireSolidBlocks && this instanceof TileRail) {
 				double floating = ((TileRail)this).percentFloating();
+				if (floating != 0) {
+					System.out.println("FLOATING " + floating);
+				}
 				if (floating > ConfigBalance.trackFloatingPercent) {
 					if (BlockRailBase.tryBreakRail(world, pos)) {
 						getWorld().destroyBlock(pos, true);
